+++ src/main/kotlin/vendor1/drink/dto/ModifyDrink.kt
@@ -1,4 +1,4 @@
-package vendor1.dto
+package vendor1.drink.dto
 
 import vendor1.drink.DrinkName
 import java.io.Serializable
