+++ src/main/kotlin/vendor1/drink/dto/RegisterDrink.kt
@@ -1,4 +1,4 @@
-package vendor1.dto
+package vendor1.drink.dto
 
 import vendor1.drink.DrinkName
 import java.io.Serializable
