+++ src/main/kotlin/vendor1/drink/processor/BuyDrinkProcessor.kt
@@ -1,6 +1,8 @@
-package vendor1.processor
+package vendor1.drink.processor
 
-import vendor1.dto.BuyDrink
+import vendor1.drink.dto.BuyDrink
+import vendor1.processor.Processor
+import vendor1.processor.entityBinder
 
 class BuyDrinkProcessor : Processor {
