+++ src/main/kotlin/vendor1/drink/processor/Command.kt
@@ -1,4 +1,4 @@
-package vendor1.processor
+package vendor1.drink.processor
 
 enum class Command {
     STATUS, REGISTER, MODIFY, SPECIFICATION, BUY, QUIT
