+++ src/main/kotlin/vendor1/drink/processor/StatusProcessor.kt
@@ -1,7 +1,9 @@
-package vendor1.processor
+package vendor1.drink.processor
 
 import vendor1.dto.Authentication
-import vendor1.dto.VendorOperation
+import vendor1.drink.dto.VendorOperation
+import vendor1.processor.Processor
+import vendor1.processor.entityBinder
 
 class StatusProcessor : Processor {
